+++ tests/net/buf/src/Makefile
@@ -1,4 +1,4 @@
-obj-y = main.o
+obj-y = main.o test_buf_context.o
 ccflags-y += -I${ZEPHYR_BASE}/subsys
 
 include $(ZEPHYR_BASE)/tests/Makefile.test
